# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DigitGuard API",
    "version": "1.0.0",
    "description": "디지털 취약 계층 보호 앱 백엔드 API"
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1", "description": "개발 서버" }
  ],
...
  "paths": {
...
    "/apps/whitelist/{userId}": {
      "get": {
        "tags": ["앱 관리"], "summary": "승인된 앱 목록",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "화이트리스트" } }
      }
    },
...
    "/devices/fcm-token": {
      "post": {
        "tags": ["디바이스"], "summary": "FCM 토큰 등록/갱신",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["userId", "fcmToken"],
          "properties": { "userId": { "type": "string" }, "fcmToken": { "type": "string" } }
        }}}},
        "responses": { "200": { "description": "등록 완료" } }
      },
      "delete": {
        "tags": ["디바이스"], "summary": "FCM 토큰 삭제",
        "requestBody": { "content": { "application/json": { "schema": {
          "type": "object", "properties": { "userId": { "type": "string" } }
        }}}},
        "responses": { "200": { "description": "삭제 완료" } }
      }
    },
    "/whitelist/{userId}": {
      "get": {
        "tags": ["화이트리스트"], "summary": "허용 앱 목록 조회",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "화이트리스트" } }
      },
      "post": {
        "tags": ["화이트리스트"], "summary": "허용 앱 추가",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["packageName"],
          "properties": { "packageName": { "type": "string" }, "appName": { "type": "string" } }
        }}}},
        "responses": { "201": { "description": "추가 완료" }, "409": { "description": "중복" } }
      }
    },
    "/whitelist/{userId}/{packageName}": {
      "delete": {
        "tags": ["화이트리스트"], "summary": "허용 앱 제거",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "packageName", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "제거 완료" }, "404": { "description": "없는 앱" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "bearerFormat": "Firebase ID Token" }
    }
  }
}